Tidy up types and use uint8_t for clarity
Fix a few oddities with the used types and replace uses of unsigned char
with uint8_t for clarity.

+++ BbcScreenWidget.h
@@ -39,7 +39,7 @@ protected:
     void paintEvent(QPaintEvent *event) Q_DECL_OVERRIDE;
 
 private:
-    void pixelCallback(int x, int y, unsigned long colour);
+    void pixelCallback(int x, int y, uint8_t colour);
 
     QImage image;
 };
